Refactor polars server: new crate bastionlab_polars client: polars client is now available under client.polars
