+++ client/src/bastionlab/__init__.py
@@ -5,6 +5,7 @@ import sys
 # Needed because internal imports within generated modules are relative
 sys.path.append(os.path.join(os.path.dirname(__file__), "pb"))
 
+# Re-exports
+
 from .client import Connection
-from .remote_polars import RemoteLazyFrame, RemoteLazyGroupBy
 from .keys import SigningKey, PublicKey
