+++ client/src/bastionlab/client.py
@@ -2,39 +2,22 @@ from dataclasses import dataclass
 import ssl
 from threading import Thread
 from time import sleep
-from typing import Any, List, TYPE_CHECKING, Optional
+from typing import Any, TYPE_CHECKING, Optional
 from hashlib import sha256
 import grpc
-
-from bastionlab.keys import SigningKey
-from grpc import StatusCode
-from bastionlab.pb.bastionlab_pb2 import (
-    ReferenceRequest,
-    Query,
-    Empty,
-)
-from bastionlab.pb.bastionlab_session_pb2 import ClientInfo
-from bastionlab.version import __version__ as app_version
-from bastionlab.pb.bastionlab_pb2_grpc import BastionLabStub
-from bastionlab.pb.bastionlab_session_pb2_grpc import SessionServiceStub
-
+from .keys import SigningKey
+from .pb.bastionlab_pb2 import Empty
+from .pb.bastionlab_pb2 import ClientInfo
+from .version import __version__ as app_version
+from .pb.bastionlab_pb2_grpc import SessionServiceStub
 import platform
 import socket
 import getpass
-import polars as pl
-from colorama import Fore
-
-from bastionlab.utils import (
-    deserialize_dataframe,
-    serialize_dataframe,
-)
-from bastionlab.policy import Policy, DEFAULT_POLICY
-from bastionlab.errors import GRPCException
 
 
 if TYPE_CHECKING:
-    from bastionlab.remote_polars import RemoteLazyFrame, FetchableLazyFrame
     from .torch import BastionLabTorch
+    from .polars import BastionLabPolars
 
 HEART_BEAT_TICK = 25 * 60
 UNAME = platform.uname()
@@ -52,8 +35,10 @@ CLIENT_INFO = ClientInfo(
 
 
 class Client:
-    __bastionlab_torch: Any = None
-    __channel: Any
+    __bastionlab_torch: "BastionLabTorch" = None
+    __bastionlab_polars: "BastionLabPolars" = None
+    __channel: grpc.Channel
+    __token: bytes
 
     def __init__(
         self,
@@ -61,8 +46,7 @@ class Client:
         token: bytes,
     ):
         self.__channel = channel
-        self.stub = BastionLabStub(channel)  # polars stub
-        self.token = token
+        self.__token = token
 
     @property
     def torch(self):
@@ -72,79 +56,13 @@ class Client:
             self.__bastionlab_torch = BastionLabTorch(self.__channel)
         return self.__bastionlab_torch
 
-    def send_df(
-        self,
-        df: pl.DataFrame,
-        policy: Policy = DEFAULT_POLICY,
-        blacklist: List[str] = [],
-    ) -> "FetchableLazyFrame":
-        from bastionlab.remote_polars import FetchableLazyFrame
+    @property
+    def polars(self):
+        if self.__bastionlab_polars is None:
+            from bastionlab.polars import BastionLabPolars
 
-        res = GRPCException.map_error(
-            lambda: self.stub.SendDataFrame(serialize_dataframe(df, policy, blacklist))
-        )
-        return FetchableLazyFrame._from_reference(self, res)
-
-    def _fetch_df(self, ref: List[str]) -> Optional[pl.DataFrame]:
-        def inner() -> bytes:
-            joined_bytes = b""
-            blocked = False
-
-            for b in self.stub.FetchDataFrame(ReferenceRequest(identifier=ref)):
-                if blocked:
-                    blocked = False
-                    print(
-                        f"{Fore.GREEN}The query has been accepted by the data owner.{Fore.WHITE}"
-                    )
-                if b.pending != "":
-                    blocked = True
-                    print(
-                        f"""{Fore.YELLOW}Warning: non privacy-preserving queries necessitate data owner's approval.
-Reason: {b.pending}
-
-A notification has been sent to the data owner. The request will be pending until the data owner accepts or denies it or until timeout seconds elapse.{Fore.WHITE}"""
-                    )
-                joined_bytes += b.data
-            return joined_bytes
-
-        try:
-            joined_bytes = GRPCException.map_error(inner)
-            return deserialize_dataframe(joined_bytes)
-        except GRPCException as e:
-            if e.code == StatusCode.PERMISSION_DENIED:
-                print(
-                    f"{Fore.RED}The query has been rejected by the data owner.{Fore.WHITE}"
-                )
-                return None
-            else:
-                raise e
-
-    def _run_query(
-        self,
-        composite_plan: str,
-    ) -> "FetchableLazyFrame":
-        from bastionlab.remote_polars import FetchableLazyFrame
-
-        res = GRPCException.map_error(
-            lambda: self.stub.RunQuery(Query(composite_plan=composite_plan))
-        )
-        return FetchableLazyFrame._from_reference(self, res)
-
-    def list_dfs(self) -> List["FetchableLazyFrame"]:
-        from bastionlab.remote_polars import FetchableLazyFrame
-
-        res = GRPCException.map_error(lambda: self.stub.ListDataFrames(Empty()).list)
-        return [FetchableLazyFrame._from_reference(self, ref) for ref in res]
-
-    def get_df(self, identifier: str) -> "FetchableLazyFrame":
-        from bastionlab.remote_polars import FetchableLazyFrame
-
-        res = GRPCException.map_error(
-            lambda: self.stub.GetDataFrameHeader(
-                ReferenceRequest(identifier=identifier)
-            )
-        )
-        return FetchableLazyFrame._from_reference(self, res)
+            self.__bastionlab_polars = BastionLabPolars(self.__channel)
+        return self.__bastionlab_polars
 
 
 class AuthPlugin(grpc.AuthMetadataPlugin):
@@ -204,8 +122,7 @@ class Connection:
             self.__exit__(None, None, None)
 
     def _heart_beat(self, stub, token):
-
-        while True:
+        while self._client is not None:
             stub.RefreshSession(Empty(), metadata=(("accesstoken-bin", token),))
             sleep(HEART_BEAT_TICK)
 
@@ -233,9 +150,14 @@ class Connection:
         self.channel = grpc.secure_channel(
             server_target, channel_cred, connection_options
         )
-        stub = BastionLabStub(self.channel)
+        stub = SessionServiceStub(self.channel)
 
-        daemon = Thread(
+        self._client = Client(
+            self.channel,
+            token,
+        )
+
+        self._daemon = Thread(
             target=self._heart_beat,
             args=(
                 stub,
@@ -244,12 +166,7 @@ class Connection:
             daemon=True,
             name="HeartBeat",
         )
-        daemon.start()
-
-        self._client = Client(
-            self.channel,
-            token,
-        )
+        self._daemon.start()
 
         return self._client
