+++ client/src/bastionlab/polars/__init__.py
@@ -1,0 +1,2 @@
+from .client import BastionLabPolars
+from .remote_polars import RemoteLazyFrame, RemoteLazyGroupBy
