+++ client/src/bastionlab/torch/__init__.py
@@ -1,4 +1,1 @@
-import os
-import sys
-
 from .client import BastionLabTorch
