+++ protos/bastionlab_polars.proto
@@ -1,0 +1,40 @@
+syntax = "proto3";
+package bastionlab_polars;
+
+message ReferenceRequest {
+    string identifier = 1;
+}
+
+message ReferenceResponse {
+    string identifier = 1;
+    string header = 2;
+}
+
+message ReferenceList {
+    repeated ReferenceResponse list = 1;
+}
+
+message SendChunk {
+    bytes data = 1;
+    string policy = 2;
+    string metadata = 3;
+}
+
+message FetchChunk {
+    bytes data = 1;
+    string pending = 2;
+}
+
+message Query {
+    string composite_plan = 1;
+}
+
+message Empty {}
+
+service PolarsService {
+    rpc SendDataFrame (stream SendChunk) returns (ReferenceResponse) {}
+    rpc RunQuery (Query) returns (ReferenceResponse) {}
+    rpc FetchDataFrame (ReferenceRequest) returns (stream FetchChunk) {}
+    rpc ListDataFrames (Empty) returns (ReferenceList) {}
+    rpc GetDataFrameHeader (ReferenceRequest) returns (ReferenceResponse) {}
+}
