--- protos/bastionlab_session.proto
@@ -1,28 +1,0 @@
-syntax = "proto3";
-package bastionlab_session;
-
-message Empty {}
-
-message ChallengeResponse {
-    bytes value = 1;
-}
-
-message SessionInfo {
-    bytes token = 1;
-}
-
-message ClientInfo {
-    string uid = 1;
-    string platform_name = 2;
-    string platform_arch = 3;
-    string platform_version = 4;
-    string platform_release = 5;
-    string user_agent = 6;
-    string user_agent_version = 7;
-}
-
-service SessionService {
-    rpc GetChallenge (Empty) returns (ChallengeResponse) {}
-    rpc CreateSession (ClientInfo) returns (SessionInfo) {}
-    rpc RefreshSession (Empty) returns (Empty) {}
-}
