+++ server/bastionlab_common/src/lib.rs
@@ -2,7 +2,8 @@ pub mod auth;
 pub mod prelude;
 pub mod session;
 pub mod telemetry;
+pub mod config;
 
 pub mod session_proto {
-    tonic::include_proto!("bastionlab_session");
+    tonic::include_proto!("bastionlab");
 }
