+++ server/bastionlab_polars/build.rs
@@ -1,0 +1,9 @@
+use std::error::Error;
+
+fn main() -> Result<(), Box<dyn Error>> {
+    println!("cargo:rerun-if-changed=build.rs");
+    println!("cargo:rerun-if-changed=../../protos/bastionlab_polars.proto");
+    tonic_build::compile_protos("../../protos/bastionlab_polars.proto")?;
+
+    Ok(())
+}
