--- server/build.rs
@@ -1,9 +1,0 @@
-use std::error::Error;
-
-fn main() -> Result<(), Box<dyn Error>> {
-    println!("cargo:rerun-if-changed=build.rs");
-    println!("cargo:rerun-if-changed=../protos/bastionlab.proto");
-    tonic_build::compile_protos("../protos/bastionlab.proto")?;
-
-    Ok(())
-}
